Tråd side + Up/down vote på forside

+++ shared/Model/Comment.cs
@@ -3,14 +3,14 @@ namespace shared.Model
 
     public class Comment
     {
-        public long CommentId { get; set; }
+        public int CommentId { get; set; }
         public DateTime Date { get; set; }
         public User User { get; set; }
         public string Text { get; set; }
         public int Votes { get; set; }
-        public long PostId { get; set; }
+        public int PostId { get; set; }
 
-        public Comment(DateTime date = new DateTime(), User user = null, string text = "", int votes = 0, long postId = 0)
+        public Comment(DateTime date = new DateTime(), User user = null, string text = "", int votes = 0, int postId = 0)
         {
             Date = date;
             User = user;
